Refactor Pic

+++ src/os/devices/usb/ehci/Ehci.cc
@@ -436,7 +436,7 @@ UsbMassStorage *Ehci::getDevice(uint32_t index) {
 void Ehci::plugin() {
     EHCI_TRACE("Assigning interrupt %d\n", pciDevice.intr);
 
-    IntDispatcher::assign(pciDevice.intr + 32, *this);
+    IntDispatcher::assign((uint8_t) pciDevice.intr + 32, *this);
     Pic::getInstance()->allow(pciDevice.intr);
 }
